Parte 2/ecommerce_etl/requirements.txt
# Requerimientos del proyecto ETL
requests==2.31.0
psycopg2-binary==2.9.7
pandas==2.0.3
PyYAML==6.0
python-dotenv==1.0.0
numpy==1.24.3